+++ yeezy_supply/__init__.py
@@ -2,14 +2,9 @@ from json import loads, JSONDecodeError
 from typing import List
 
 from jsonpath2 import Path
-from lxml import etree
-from getpass import  getpass
-import time
-import cfscrape
-from requests.exceptions import ReadTimeout
 from requests import get
+from requests.exceptions import ReadTimeout
 from user_agent import generate_user_agent
-
 
 from core import api
 from core.api import IndexType, TargetType, StatusType
@@ -21,7 +16,7 @@ class Parser(api.Parser):
         super().__init__(name, log)
         self.catalog: str = 'https://www.yeezysupply.com/api/yeezysupply/products/bloom'
         self.user_agent = generate_user_agent()
-        self.interval: float = 1
+        self.interval: int = 1
 
     def index(self) -> IndexType:
         return api.IInterval(self.name, 120)
@@ -30,20 +25,24 @@ class Parser(api.Parser):
         while True:
             try:
                 return [
-                    api.TInterval(i.current_value['product_name'],
-                    self.name,
-                    'https://www.yeezysupply.com/api/products/' + i.current_value['product_id'],
-                    self.interval)
+                    api.TInterval(
+                        i.current_value['product_name'],
+                        self.name,
+                        'https://www.yeezysupply.com/api/products/' + i.current_value['product_id'],
+                        self.interval
+                    )
                     for i in Path.parse_str('$[*]').match(
-                        loads(get(self.catalog, headers={'user-agent': self.user_agent,
-                                                         'referer': 'https://www.yeezysupply.com',
-                                                         'connection': 'keep-alive',
-                                                         'cache-control': 'max-age=0', 'upgrade-insecure-requests': '1',
-                                                         'sec-fetch-dest': 'document',
-                                                         'accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8',
-                                                         'sec-fetch-site': 'same-origin', 'sec-fetch-mode': 'navigate', 'sec-fetch-user': '?1',
-                                                         'accept-language': 'en-US,en;q=0.9',
-                                                         }, timeout=3).text)
+                        loads(get(self.catalog, headers={
+                            'user-agent': self.user_agent,
+                            'referer': 'https://www.yeezysupply.com',
+                            'connection': 'keep-alive',
+                            'cache-control': 'max-age=0', 'upgrade-insecure-requests': '1',
+                            'sec-fetch-dest': 'document',
+                            'accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8',
+                            'sec-fetch-site': 'same-origin', 'sec-fetch-mode': 'navigate',
+                            'sec-fetch-user': '?1',
+                            'accept-language': 'en-US,en;q=0.9',
+                        }, timeout=3).text)
                     )
                 ]
             except ReadTimeout:
@@ -56,14 +55,16 @@ class Parser(api.Parser):
                     try:
                         sizes_json = Path.parse_str('$.variation_list[*]').match(loads(
                             get(f'https://www.yeezysupply.com/api/products/{target.data.split("/")[5]}/availability',
-                                headers={'user-agent': self.user_agent,
-                                         'referer': f'https://www.yeezysupply.com/product/{target.data.split("/")[5]}',
-                                         'connection': 'keep-alive', 'cache-control': 'max-age=0',
-                                         'upgrade-insecure-requests': '1', 'sec-fetch-dest': 'document',
-                                         'accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8',
-                                         'sec-fetch-site': 'same-origin', 'sec-fetch-mode': 'navigate',
-                                         'sec-fetch-user': '?1',
-                                         'accept-language': 'en-US,en;q=0.9'}, timeout=3).text))
+                                headers={
+                                    'user-agent': self.user_agent,
+                                    'referer': f'https://www.yeezysupply.com/product/{target.data.split("/")[5]}',
+                                    'connection': 'keep-alive', 'cache-control': 'max-age=0',
+                                    'upgrade-insecure-requests': '1', 'sec-fetch-dest': 'document',
+                                    'accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8',
+                                    'sec-fetch-site': 'same-origin', 'sec-fetch-mode': 'navigate',
+                                    'sec-fetch-user': '?1',
+                                    'accept-language': 'en-US,en;q=0.9'
+                                }, timeout=3).text))
                         break
                     except ReadTimeout:
                         pass
@@ -74,15 +75,17 @@ class Parser(api.Parser):
                         content: dict = loads(
                             get(
                                 target.data,
-                                headers={'user-agent': generate_user_agent(),
-                                         'referer': 'https://www.yeezysupply.com',
-                                         'connection': 'keep-alive', 'cache-control': 'max-age=0',
-                                         'upgrade-insecure-requests': '1', 'sec-fetch-dest': 'document',
-                                         'accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8',
-                                         'sec-fetch-site': 'same-origin', 'sec-fetch-mode': 'navigate',
-                                         'sec-fetch-user': '?1',
-                                         'accept-language': 'en-US,en;q=0.9'
-                                         }, timeout=3
+                                headers={
+                                    'user-agent': generate_user_agent(),
+                                    'referer': 'https://www.yeezysupply.com',
+                                    'connection': 'keep-alive', 'cache-control': 'max-age=0',
+                                    'upgrade-insecure-requests': '1', 'sec-fetch-dest': 'document',
+                                    'accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8',
+                                    'sec-fetch-site': 'same-origin', 'sec-fetch-mode': 'navigate',
+                                    'sec-fetch-user': '?1',
+                                    'accept-language': 'en-US,en;q=0.9'
+                                }
+                                , timeout=3
                             ).text
                         )
                         break
